# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Post:
    properties:
      author_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      reply_to:
        type: integer
      topic_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  postrequests.UpdateRequest:
    properties:
      content:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  response.IDResponse:
    properties:
      id:
        example: 123
        type: integer
    type: object
  response.PostsResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/entity.Post'
        type: array
    type: object
  response.SuccessMessageResponse:
    properties:
      message:
        example: operation was successful
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API for forum service
  title: Forum Service API
  version: "1.0"
paths:
...
  /posts/{id}:
    delete:
      description: Deletes a post by its ID. Requires authentication and ownership
        or admin role.
      parameters:
      - description: Post ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Post deleted successfully
          schema:
            $ref: '#/definitions/response.SuccessMessageResponse'
        "400":
          description: Invalid post ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an owner or admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a post
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Updates a post. Requires authentication and ownership or admin
        role.
      parameters:
      - description: Post ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Post update data (only content)
        in: body
        name: post_update
        required: true
        schema:
          $ref: '#/definitions/postrequests.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            $ref: '#/definitions/response.SuccessMessageResponse'
        "400":
          description: Invalid post ID or request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an owner or admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a post
      tags:
      - posts
...
  /topics/{id}/posts:
    get:
      description: Retrieves a list of posts for a topic ID.
      parameters:
      - description: Topic ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved posts
          schema:
            $ref: '#/definitions/response.PostsResponse'
        "400":
          description: Invalid topic ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Topic not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get posts by topic ID
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Creates a new post in atopic. Requires authentication.
      parameters:
      - description: Topic ID to create post in
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Post data to create. ID, TopicID, AuthorID, Username, CreatedAt,
          UpdatedAt will be ignored or overridden.
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/entity.Post'
      produces:
      - application/json
      responses:
        "200":
          description: Post created successfully
          schema:
            $ref: '#/definitions/response.IDResponse'
        "400":
          description: Invalid topic ID or request payload, or topic not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not authorized)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new post in a topic
      tags:
      - posts
swagger: "2.0"